pyproject.toml
[project]
name = "kairosium-finance-anomaly-agent"
version = "0.1.0"
description = "Pipeline multi-agents de détection d'anomalies financières NovaPay (ADK + GCP)"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.60.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-monitoring>=2.0.0",
    "google-cloud-firestore>=2.0.0",
    "opentelemetry-sdk>=1.20.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["orchestrator", "ingestion_agent", "scoring_agent", "shared", "config"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
